+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 - Added the `prefixGqlRootTypes` config setting, which determines whether the `gqlTypePrefix` config setting should impact `query`, `mutation`, and `subscription` types. ([#7552](https://github.com/craftcms/cms/issues/7552))
 - Added `craft\helpers\Html::failMessageInput()`.
 - Added `craft\helpers\Html::successMessageInput()`.
+- Added `craft\services\Gc::deletePartialElements()`.
 
 ### Changed
 - It’s now possible to delete entries from the “All entries” source. ([#7545](https://github.com/craftcms/cms/issues/7545))
 - Garbage collection now deletes orphaned rows in the `drafts` and `revisions` tables.
+- Garbage collection now deletes incomplete element data, for built-in element types. ([#6434](https://github.com/craftcms/cms/issues/6434))
 
 ### Fixed
 - Fixed a bug where Craft wasn’t deleting unpublished drafts when converting a Channel or Structure section to a Single. ([#37542](https://github.com/craftcms/cms/issues/7542))